Refactor to have a superclass for Notify renewal letters
