+++ app/services/notify_ad_renewal_letter_service.rb
@@ -1,20 +1,6 @@
 # frozen_string_literal: true
 
-require "notifications/client"
-
-class NotifyAdRenewalLetterService < ::WasteCarriersEngine::BaseService
-  # So we can use displayable_address()
-  include ::WasteCarriersEngine::ApplicationHelper
-
-  def run(registration:)
-    @registration = NotifyRenewalLetterPresenter.new(registration)
-
-    client = Notifications::Client.new(WasteCarriersEngine.configuration.notify_api_key)
-
-    client.send_letter(template_id: template,
-                       personalisation: personalisation)
-  end
-
+class NotifyAdRenewalLetterService < NotifyRenewalLetterService
   private
 
   def template
@@ -30,21 +16,5 @@ class NotifyAdRenewalLetterService < ::WasteCarriersEngine::BaseService
       renewal_cost: @registration.renewal_cost,
       renewal_url: @registration.renewal_url
     }.merge(address_lines)
-  end
-
-  def address_lines
-    address_values = [
-      @registration.contact_name,
-      displayable_address(@registration.contact_address)
-    ].flatten
-
-    address_hash = {}
-
-    address_values.each_with_index do |value, index|
-      line_number = index + 1
-      address_hash["address_line_#{line_number}"] = value
-    end
-
-    address_hash
   end
 end
